refactor timer
